Added the get_repeats_sam method. This method is a fairly top_level method that returns the rmsk dictionary with all the repeats appended to the value list. It calls upon the sort_sam_list and valid_sam repeats methods.
Added the sort_sam submethod to the parse_rmsk method. This method creates a list of all the sam chromosome reads' positions and sorts them in alphanumerical order. Uses the merge sort algorithm

Added the valid_sam_repeats method. This method basically takes in the sorted sam file to find which sam reads overlap with the target rmsk subfamily. This method calls upon the method search_sam to scan through the sorted sam file to find the necessary sam reads

Added the search_sam method. This method uses a binary search algorithm to isolate the sam reads that overlap with the rmsk subfamily sequence.