Finished the method format_appended_sam, which aligns the relevant sam reads to the reference sequence in the rmsk file.
Created the method calculate_methylation_levels, which calculates the methylation score for each cytoside in the consensus sequence in the rmsk file by averaging across all the sam reads.

Created the method update_methylation_dictionary, which updates the methylation dictionary based on the scores calculated in the calculate_methylation_levels.